Extracted replace-permalink to separate module
